Add files via upload
The notebook above was created using Google Colab and contains code specific to that IDE. The code in reference is a snippet that enables colab to access a device's webcam in order to execute Detectron2's panoptic segmentation. To ensure this notebook executes, verify that the Runtime type has GPU selected as the "Hardware Accelerator". The output file, containing the segmented video will be availabe in the contents folder as "out.mp4" after running the notebook.